+++ extensions/google-drive/src/types.ts
@@ -1,0 +1,15 @@
+export type FileInfo = {
+  name: string;
+  path: string;
+  displayPath: string;
+  fileSizeFormatted: string;
+  createdAt: Date;
+  updatedAt: Date;
+  favorite: boolean;
+  searchScore?: number;
+};
+
+export type Preferences = {
+  shouldShowHiddenFiles: boolean;
+  googleDriveRootPath: string;
+};